requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
WTForms==2.3.3
psycopg2-binary==2.9.1
google-cloud-storage==1.42.0
gunicorn==20.1.0